Update to temporary file name handling
I made a few changes to the way the temporary filename handling was done to be more portable and adherent to general conventions.

First of all, I removed the MAX_TEMPFILE_NAME macro and used NAME_MAX from limits.h. This is usually 255 on POSIX systems.

In addition to that, instead of using '/tmp/' in the filename, I used snprintf() to concatenate the name template onto value of P_tmpdir in case the system is set to use a different temporary directory than '/tmp'.  Therefore I made the 'tempfile_name' array to be the size of PATH_MAX + NAME_MAX.

Finally, since most programs use the template 'programname.XXXXXX' I changed the temporary file name template to 'imgcat.XXXXXX'